+++ src/frontend/web_application/src/scenes/ContactBook/presenter.jsx
@@ -14,7 +14,8 @@ import {
   withTags, TagsForm, getCleanedTagCollection, getTagNamesInCommon,
 } from '../../modules/tags';
 import TagList from './components/TagList';
-import ImportContactButton from './components/ImportContactButton';
+// XXX : Disabled until fixed on backend.
+// import ImportContactButton from './components/ImportContactButton';
 import { withTagSearched } from './hoc/withTagSearched';
 import './style.scss';
 import './contact-book-menu.scss';
@@ -354,12 +355,15 @@ class ContactBook extends Component {
                       <Trans id="contact-book.action.add">Add</Trans>
                     </Button>
                   </NavItem>
-                  <NavItem>
-                    <ImportContactButton
-                      className="s-contact-book__action-button"
-                      onUploadSuccess={this.handleUploadSuccess}
-                    />
-                  </NavItem>
+                  {
+                    // XXX : Disabled until fixed on backend.
+                    /* <NavItem>
+                      <ImportContactButton
+                        className="s-contact-book__action-button"
+                        onUploadSuccess={this.handleUploadSuccess}
+                      />
+                    </NavItem> */
+                  }
                 </NavList>
               </div>
               <div>