+++ src/frontend/web_application/src/modules/draftMessage/components/IdentitySelector/presenter.jsx
@@ -64,7 +64,7 @@ class IdentitySelector extends Component {
                   {' '}
                   &lt;
                   {identity.identifier}
-&gt;
+                  &gt;
                 </Fragment>
               ),
               value: identity.identity_id,