+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Simple detection of PGP inline message
+- Apiv2 create and delete contacts does not use user.shard_id
+- Better logging for apiv2 and mq-worker
+- Do not fail if ContactLookup raise a NotFound
+
 ## [0.18.2] 2019-04-26
 
 ### Fixed